AudioXtreamer: - this program is now the owner of the usb client and uses a memory mapped file to interchange the audio buffers with the asio driver. Named mutexes for synchronization. - There is now a small property sheet dialog as main window to change the asio settings and visualize basic status information.
TortugASIO:
- All usb driver functionality moved out to AudioXtreamer and the asio driver now only communicates with it through memory mapped file, named mutexes and sendmessage
- it is now possible to have a x64 dll to communicate with a  x86 Xtreamer and viceversa.

+++ TT_AUFI/UsbDK/UsbDKWrap.h
@@ -183,9 +183,9 @@ extern struct usbdk_lib UsbDk;
 bool load_usbdk();
 void unload_usbdk();
 
-int control_transfer(HANDLE handle, uint8_t bmRequestType, uint8_t bRequest, uint16_t wValue, uint16_t wIndex,
+int64_t control_transfer(HANDLE handle, uint8_t bmRequestType, uint8_t bRequest, uint16_t wValue, uint16_t wIndex,
   unsigned char *data, uint16_t wLength, unsigned int timeout);
 
-int bulk_transfer(HANDLE handle,
+int64_t bulk_transfer(HANDLE handle,
   uint8_t endpoint, uint8_t *data, int length,
   int *actual_length, OVERLAPPED * ovlp, unsigned int timeout);
